requirements.txt
pyTelegramBotAPI==4.6.1
requests==2.28.1
boto3
python-decouple==3.4
